Finished adding faq
There is a missing link for question 13 that can be added once we have
it somewhere on the page.

@theo-armour we added the FAQ with answers to the about page and removed
the .md file. Let us know your thoughts on the questions and answers as
well as links. Will be happy to provide more links if there is any
missing from the page.
www.ladybyg.tools/eval/about#faq