Ladybug Tools - Wikipedia Article Notes * Add details tags to the major sections of the document * Add links and notes to: Websites standards we want to to meet or exceed
When I started thinking about this project what I has in mind was an article about a library or framework.
Well, that does not really seem to be a thing in Wikipedia.
Writing about Ladybug Tools as a single app seems not to work as well.
So my current thinking is to write about LT as a FOSS organization.